Rt1 field validation

+++ src/main/java/com/example/lab5/Controllers/TvController.java
@@ -1,18 +1,15 @@
-package com.example.lab5;
+package com.example.lab5.Controllers;
 
 
+import com.example.lab5.Repositories.Entities.TvEntity;
+import com.example.lab5.Services.TvService;
+import jakarta.validation.Valid;
 import lombok.NonNull;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
-
-import java.util.ArrayList;
-import java.util.Collection;
-
-import static org.springframework.data.jpa.domain.AbstractPersistable_.id;
 
 
 @RequiredArgsConstructor
@@ -24,16 +21,16 @@ public class TvController {
 
     @GetMapping("/tv")
     public String getAllTv(@ModelAttribute TvEntity tvEntity, Model model){
-        model.addAttribute("tvEntities", tvService.getAllTv());
+        model.addAttribute("tvEntities",  tvService.getAllTv());
         return "tv";
     }
 
-
-
     @PostMapping("/add")
-    public String addTv(@ModelAttribute TvEntity tvEntity, Model model){
+    public String addTv(   @ModelAttribute  @Valid TvEntity tvEntity, BindingResult bindingResult){
 
-        model.addAttribute("tvEntity", tvEntity);
+        if (bindingResult.hasErrors()) {
+            return "add";
+        }
         tvService.addTv(tvEntity);
         return "addResult";
     }
@@ -45,8 +42,14 @@ public class TvController {
     }
 
     @PostMapping("/delete")
-    public String deleteTv(@ModelAttribute TvEntity tvEntity, Model model){
-         model.addAttribute("tvId", tvEntity);
+    public String deleteTv(@ModelAttribute  TvEntity tvEntity, BindingResult bindingResult, Model model){
+
+        @Valid int id = tvEntity.getId();
+        model.addAttribute("tvId", id);
+        if (bindingResult.hasErrors()) {
+            return "delete";
+        }
+
          if (tvService.existsTv(tvEntity.getId())) {
              tvService.deleteTv(tvEntity.getId());
              return "deleteResult";
@@ -56,7 +59,7 @@ public class TvController {
 
     @GetMapping("/delete")
     public String deleteTvResult(Model model){
-        model.addAttribute("tvId", new TvEntity());
+        model.addAttribute("tvEntity", new TvEntity());
         return "delete";
     }
 
@@ -69,7 +72,11 @@ public class TvController {
     }
 
     @PostMapping("/edit")
-    public String editTv(@ModelAttribute TvEntity tvEntity, Model model){
+    public String editTv(@ModelAttribute  TvEntity tvEntity,BindingResult bindingResult ,Model model){
+        model.addAttribute("tvId", tvEntity.getId());
+        if (bindingResult.hasErrors()) {
+            return "checkConditionEdit";
+        }
         if (tvService.existsTv(tvEntity.getId())) {
             model.addAttribute("tvEntity", tvService.loadUserById(tvEntity.getId()));
             return "editExist";
@@ -89,7 +96,12 @@ public class TvController {
     }
 
     @PostMapping("/search")
-    public String searchTv(@ModelAttribute TvEntity tvEntity, Model model){
+    public String searchTv(@ModelAttribute  TvEntity tvEntity, BindingResult bindingResult, Model model){
+        @Valid int id = tvEntity.getId();
+        model.addAttribute("tvId", id);
+        if (bindingResult.hasErrors()){
+            return "search";
+        }
         if (tvService.existsTv(tvEntity.getId())) {
             model.addAttribute("tvEntity", tvService.loadUserById(tvEntity.getId()));
             return "searchResult";
@@ -105,11 +117,19 @@ public class TvController {
     }
 
     @PostMapping("/search/by/price")
-    public String searchTvByLowPrice(@ModelAttribute TvEntity tvEntity, Model model){
+    public String searchTvByLowPrice(@ModelAttribute  TvEntity tvEntity, BindingResult bindingResult, Model model){
+        @Valid double price = tvEntity.getPrice();
+        model.addAttribute("tvPrice", price);
+        if (bindingResult.hasErrors()){
+            return "searchByPrice";
+        }
+
+        model.addAttribute("tvEntities", tvService.searchByLowPrice(tvEntity.getPrice()));
         if (tvService.searchByLowPrice(tvEntity.getPrice()).isEmpty()){
+
             return "searchByPriceNoResult";
         }
-        model.addAttribute("tvEntity", tvService.searchByLowPrice(tvEntity.getPrice()));
+
         return "searchByPriceResult";
     }
